Add Event Pump; Closes #9
Problem
-------

There is no central mechanism to link animation and simulation. There is no
central loop that can drive the simulation library.

Solution
--------

Create a service which produces a singleton that executes a set of callbacks on
each animation frame. This provides a consistent mechanism for the simulation
and controls to interact with the flow of the program.

Areas of Impact
---------------

This refactors the file structure of the program. This will cause grief for any
pending progress. We can hash out how to merge this.

Howto Test
----------

- [ ] Open the application
- [ ] Notice the FPS counter directive showing the current state and FPS
- [ ] Click the pause and resume to interact with the eventPump service
- [ ] Notice the state and FPS change with the eventPump state.

References
----------

- /cc @mkinsey @aaroncameron21
- [GH Issue](https://github.com/orbitable/bridge/issues/9)