Dockerize Bride for Deployment
Problem
=======

There no means to automatically build the project for deployment.

Solution
========

Add configuration so that an automated docker build can occur whenever there is
push to a branch or PR. This allows the demo service to reload any new additions
to the project more frequently.

This containerization stuff is kind of ancillary. You don't have to worry so
much about it. Just know it is there.

Howto Test
==========

- [ ] Confirm the [automated docker service][docker] has built a container

[docker]: https://hub.docker.com/r/orbitable/bridge/builds/