Index paths by body ID
Problem
-------

Paths are stored by an index value. This breaks as the number of bodies in a
simulation changes and paths are now aligned with a different body.

Solution
--------

Use an object which maps a body's id from Engine to a collection of positions.

How To Test
-----------

The application should have no regressions.