Added Solid UX Around Logins
Problem
-------

Failures when dealing with registration and logins were unreported to the user
and resulted in poor user experience. A user had no idea of what was happening
in the background and did not know about failures when bad information was
submitted and the back end server returned a failed response.

Solution
--------

Implement a minor state machine within the registration and sign in modals that
directs the user as the registration and login processes are completed with the
back end server. The modal indicates state changes to the user and dismisses on
success. A user has an opportunity to resubmit information if bad data was
supplied.