Refactor Simulation Rendering
Problem
-------

The vector arrows render based of a transient data structure which causes them
to redraw multiple times. The vector arrow is a function of a given body and
should only draw once in accordance with the properties of that body.

Solution
--------

Refactor how vector arrows are drawn to use the simulation body as the d3 data
source. This ensures the arrows are drawn once per body.