Add Saving/Loading Simulation From Mission Control
Problem
-------

It is not possible to load or save a simulation to mission control form the
bridge interface.

Solution
--------

Tie the Simulation and UserSimulation resource services to the UI to to get
basic functionality. This allow the admin panel to save a current simulation
state as well as updates the app router to load a simulation state for a given
ID.