Correct Dragging of Habitable Zones
Problem
-------

A drag event entered on the habitable zone causes the center body to be
reposition with the drag. This is incorrect as the habitable zone is a function
of a body's luminosity and position. It should move with the body exclusively.

Solution
--------

The call on the d3 selection of all circles cause the body drag handle to be
bound to the habitable zone, `d3.selectAll('circle').call(drag);`. This is the
cause of the issue. The body positioning drag events should be bound to the
body as exemplified below.

**Bonus** I restructured the code to follow the Google JS coding styles as
recommended by JSCS. Enable the JSCS linter in your editor to get hints for a
concise code style.