Add Correct Reset Behaviour
Problem
-------

When a user clicks reset within bridge a random simulation is generated. The
correct behavior is to reset the currently selected simulation to it's initial
state.

Solution
--------

Check the route parameters for a simulation ID and reload the simulation with
the given ID. If none is present, load the random simulation.