Nest NoteController Correctly
Problem
-------

The noteController was added above the form elements which does not encapsulate
the note data panel heading. The close button included in this heading was not
able to reference the note controller preventing it from closing the panel on
click.

Solution
--------

Restructure the panel data body to so the controller encompasses the whole panel
structure.