Use selectedNote Scope Field for Note Selection
Problem
-------

When selecting a note the selected body scope object is used to hold it's
reference but this causes issues when dealing with interacting with the selected
body as if it is a body.

Solution
--------

Add an independent scope field labeled `selectedNote` which holds the handle to
the last selected note so there is no conflict between the two.