Flag Scope as Dirty On Selection
Problem
-------

When updating the scope object there is not explicit call to $apply which
prevents angular from diffing the scope update.

Solution
--------

Call apply when selecting a new body or note to signal to angular that the scope
has been updated. This causes any other reference to the scope to be
reevaluated.