Add Units Service and Directive
Problem
-------

The user selectable units are non functional as the model for body data has
beens shifted.

Solution
--------

A unit directive allows a for element to parse and cast a raw model data value
into a derived value based on a selected unit set in the Units service. This
allows the form to show the value of the simulation in different units without
haveing to convert within the simulator.