Unify Simulation Reset
Problem
-------

When a new simulation is created the existing paths are not wiped causing the
exisiting paths to be reattributed to the new bodies causing strange path
issues.

Solution
--------

Write a single handler for simulation loading and use it in both new simualation
as well as resets agains the end. This ensures the logic lives in one place and
changes to one will affect all applicable instances.