Update Scope with Zoom or Pan Event
Problem
-------

Some objects which depend on scope value changes for positioning info are not
being redrawn when the canvas is panned or zoomed.

Solution
--------

The scope is never updated in these events causing angular to never re render
anything. By forcing an apply on top level pans and zooms all of the dependent
objects are updated.