Add Reset Interceptors for Path Resetting
Problem
-------

Path data can be left in strange states which leaves path artifacts on the
screen.

Solution
--------

Remove paths from the global namespace and add them as a service. In additon
interceptors around the simulation reset and resetLocal calls allow the path
data to be reset in cordination with these function calls. This ensures the path
data is reset any time the simulation bodies move outside of a update call.